+++ examples/demos/components/index.ts
@@ -1,3 +1,5 @@
 export { MapRender as control } from './control';
+export { MapRender as layerPopup } from './layerPopup';
 export { MapRender as marker } from './marker';
+export { MapRender as popup } from './popup';
 export { MapRender as swipe } from './swipe';