+++ packages/renderer/CHANGELOG.md
@@ -1,12 +1,22 @@
 # Change Log
 
-## 2.21.9-beta.0
+## 2.21.10
 
 ### Patch Changes
 
-- Updated dependencies [[`0913661`](https://github.com/antvis/L7/commit/09136611f4521c5ecb892a7ddf67a8e680e74429)]:
-  - @antv/l7-core@2.21.9-beta.0
-  - @antv/l7-utils@2.21.9-beta.0
+- [#2466](https://github.com/antvis/L7/pull/2466) [`4962f65`](https://github.com/antvis/L7/commit/4962f65a27cbd4a1becb09a5c6bab89368099fca) Thanks [@lvisei](https://github.com/lvisei)! - fix: use antialias option in webgl
+
+- Updated dependencies []:
+  - @antv/l7-core@2.21.10
+  - @antv/l7-utils@2.21.10
+
+## 2.21.9
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @antv/l7-core@2.21.9
+  - @antv/l7-utils@2.21.9
 
 ## 2.21.8