+++ packages/three/CHANGELOG.md
@@ -1,11 +1,18 @@
 # @antv/l7-three
 
-## 2.21.9-beta.0
+## 2.21.10
 
 ### Patch Changes
 
 - Updated dependencies []:
-  - @antv/l7@2.21.9-beta.0
+  - @antv/l7@2.21.10
+
+## 2.21.9
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @antv/l7@2.21.9
 
 ## 2.21.8